Bulk update resilience: pruning safety, idempotent diffs, notification grouping
Extract diff logic into WatcherDiffRunner with progressive lastDiffedSnapshotId updates and idempotent report insertion via sourceSnapshotId column. SnapshotWorker now runs diffs inline and aggressively prunes to 2 snapshots after success, while keeping a hard cap of 20 during burst saves. Notification grouping with summary for multi-app changes. Simplified PermissionWatcherWorker to thin delegate.
